+++ js/video/src/Video.svelte
@@ -18,14 +18,16 @@
 	export let include_audio: boolean;
 
 	const dispatch = createEventDispatcher<{
-		change: FileData | null;
+		change: any;
 		clear: undefined;
 		play: undefined;
 		pause: undefined;
-		ended: undefined;
+		end: undefined;
 		drag: boolean;
 		error: string;
 		upload: FileData;
+		start_recording: undefined;
+		stop_recording: undefined;
 	}>();
 
 	function handle_load({ detail }: CustomEvent<FileData | null>) {
@@ -57,6 +59,8 @@
 			mode="video"
 			on:error
 			on:capture={({ detail }) => dispatch("change", detail)}
+			on:start_recording
+			on:stop_recording
 		/>
 	{/if}
 {:else}
@@ -68,7 +72,8 @@
 			subtitle={subtitle?.data}
 			on:play
 			on:pause
-			on:ended
+			on:stop
+			on:end
 			mirror={mirror_webcam && source === "webcam"}
 		/>
 	{:else if value.size}